app/requirements.txt
attrdict==2.0.1
glm==0.4.3
guided_filter_pytorch==3.7.5
imageio==2.9.0
kornia==0.5.0
lxml==4.6.1
mapbox_earcut==1.0.0
matplotlib==3.2.2
Numeric==24.2
numpy==1.19.2
open3d==0.12.0
opencv_python==4.5.1.48
Pillow==9.1.0
PyGLM==1.99.0
pylsd==0.0.2
pymesh==1.0.2
PyOpenGL==3.1.6
PyQt5==5.15.6
pytorch3d==0.3.0
PyYAML==6.0
scikit_image==0.17.2
scipy==1.2.0
sh==1.14.1
Shapely==1.7.0
skimage==0.0
sympy==1.6.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.50.2
